requirements.txt
Django>=4.1.0,<4.2.0
djangorestframework>=3.14.0,<3.15
psycopg2>=2.9.5,<2.10
drf-spectacular>=0.24.2,<0.25
drf-nested-routers>=0.93.4,<0.94
Pillow>=9.3.0,<9.4
